requirements.txt
flask
flask-cors
gunicorn
python-dotenv
azure-ai-inference==1.0.0b9